requirements_rendering.txt
mitsuba
matplotlib
drjit
tqdm
albumentations
opencv-python
hsluv
numpy
pillow